requirements.txt
mozilla-django-oidc==3.0.0
psycopg2-binary==2.9.6
Django==4.2.3
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
